restructure the cogs
splits the Database cog in two, WikiDatabase and PermissionsDatabase, the former being exclusively for
page manipulation functionality.

This is for the upcoming switch to having sql commands in separate files.
It made more sense to have two separate dicts for the wiki queries and the permissions queries
than it did to make them all in one dict. (You'll see in the next commit)