Use better paginators
For most of the commands I want a paginator that limits the size of a page by the number of entries
contained in it, not by the amount of text it has. For that I switched to R. Danny's Pages class.

For consistency, I also switched the text-based pagination from Jishaku to TextPages.